Update LAB_13_Azure Monitor.md
Hi Microsoft and Guys, somehow this GitHub idea of updating the official documentation is far from good. Many time ago the Blue Books MOC updated by Microsoft were not perfect but good. Prepared by Microsoft. When everyone can contribute even if do not have clear idea about particular area it doesnt work well. Someone named this chapter 13 - Azure Monitor, when in fact it has nothiong to do with the Azure Monitor. Azure Monitor is a tool in Azure for monitoring. Maybe someone do not know that. This chapter do not describe anything related to the Azure Monitor. Instead it describes of how to configure the "Azure Log Analytics Workspace" for VMs. Nothing more. It should be named "Azure Log Analytics" not "Azure Monitor" (Azure Portal -> Monitor). I'm sorry if someone feels not comfortable with my comment but come on. This is an official Microsoft training material. Microsoft trainings costs approx 1000$ each per person. It's not a freeware open source and personal GitHub initiative. For that some level of quality and consistency should be kept. Especially that people use that to prepare for exams and daily work.